Add Macro preprocessor
Add a FakeMacroDriver that provides a simple environment to compile Pep10asm files.
Create a Macro Assembler that dispatches assembly task to different parts of the toolchain.
Add a pathing algorithm for graphs.

Create a preprocessor that detects macros, validates macro invocation syntax, check macro semantics, as well as detecting and recovering from circular macro inclusion.
Add stub to main that triggers all of this code.