Remove appendSourceLineMethod from AsmCode.
This method was confusingly named.
It did not generate the source form of a line of code.
Instead, it generated the listing for that line of code.
We already have functions to generate listing and source as a QString.
Therefore this method is redundant at best, and confusing at worst.