[PRE & BUILD] Handle Macro declaration in preprocessor.
The first line of a macro file declares the macro name and argument count.
The preprocessor skipped over this line of code, and left its parsing up to the assembler.
This required later processing stages (linker, annotator) to be aware of the implementation of the macro facility.
Therefore, this line should be stripped from the macro text by the preprocessor, and verified at this point in time.
Now the assembler is not required to skip the first line of input on macro modules.