[LINK] Account for BURNS during OS linking.
Implemented shiftForBURN(...), which properly relocates all code lines in operating system so that the last address is the argument of the burn.
It also disables object code generation for any instruction preceding the BURN.
If the forceBurn0xFFFF is set, and the BURN has an argument other than 0xFFFF an error will be generated indicating this mistake.
ALIGN directives before the burn will generate the correct number of bytes.

Updated FakeMacroDriver to assemble real operating system.