Make numPrint take argument via stack.
numPrint originally was baked in to DECO, meaning it had access to DECO's locals.
Now, numPrint is a free function that is called from multiple places in the OS.
It expected that the number to print was residing in A, which was a non-standard calling convention for Pep10.
Now, an extra entry on the stack contains the number to operate on.