[OS] Add additional protection to trap handler
Switch from mutex guards in OS to a trap handler lookup table.
This approach guarantees that only allowed addresses may be used as a T argument for U/SCALL.
This prevents several bugs involving recursion in the operating system, and makes failures easier to diagnose for students.
It comes at an average cost of 20 cycles / system call, peaking at 65 / system call for items at the end of a list.