Simplify global allocations tracking.
Reduce code replication in the generation of object code for BLOCK, BYTE, WORD by merging the parse*(...) methods.
These methods' bodies have been inserted into the resolveGlobals() function, as I was able to identify common components.
The result is that BYTE is now more versatile - it may declare arrays and structs - so long as they are of size 1.