Create error class to handle error lifetimes.
Handling error messages has become complex with the addition of warnings.
Properly mapping an error message from the module in which it resides to the source pane or assembler listing  is also difficult.
Therefore, a class that encapsulates this information (AErrorMessage) will ease error propagation.

Each error message indicates which module instance caused the error, and which source line within the module raised the error.
This combination of instance index and source line is unique, allowing for for easier mapping from error-to-screen.

Add a ErrorDictionary that encapsulates adding and finding error messages.