Utilize error dictionary.
Instead of using the list of <int, String> pairs for communicating error messages, switch to the more powerful ErrorDictionary.
This allows tracking of which modules produced the error, as well as mapping errors into the root module.
Mapping errors into the root module as part of the build step makes appending the error messages to source lines easier.

Extensive unit testing of the frontend and linker seems to indicate that programs known to be bad do indeed fail.
More importantly, the error messages mapped to the line which caused the error.