Add unit testing functionality to Pep10 project.
Due to the comprehensive nature of the changes made to the assembler, I felt it necessary to add automated unit tests to the Pep10 program.
For now, I only unit test the preprocessor, but eventually I will check each stage of the frontend and backend with extensive unit testing for failure.

To unit test for "successes" I will run every program in the figures ending in .pep through the assembler, and validate that is assembles properly.