Fix decimal constant, address mode parsing.
Fix a bug in the addressing mode parser where it would match any instance of an adressing mode character.
This would allow it to search into comments for these characters, causing strange syntax errors.

Fixed a similar issue in the dec constant handler that allowed for ++, -- and other strange combinations.