Add token buffer unit tests.
Add unit tests for the the token buffer, which provides tools for matching streams of tokens at once.
This functionality builds on top of the tokenizer, which is only designed to recognize a single token at once.
Many unit tests that "should" have worked in the tokenizer failed because they operated on a stream of tokens rather than an individual token.
So, these failing tests have been moved to the TokenBufferTest class.