Add test cases foe missing END.
Test various combinations of missing .END sentinels.
Check that an empty program (user or OS) fails to assemble because of a missing .END.
Check that a non-empty program (user or OS) fails to assemble because of a missing .END.
Check that when a program includes a macro that is missing a .END, the program fails to compile as a whole.