[PREPROCESS] Fix preprocessor to pass unit tests.
Fix bug where preprocessor would incorrectly match macro identifiers.
For example, @44asra4 would be matched even though a macro may not start with a number.
However, the preprocessor had detected an identifier at some point in the line, even if it was non-contiguous with the @.
This has been rectified by checking the start of a match, and additional unit tests have been added to detect if this error re-occurs.