Update tokenizer tests to reflect discussion.
As per my discussion with Dr. Warford, the tokenizer should do as little as possible.
Therefore, if a malformed dot command is generated (one with a number in it), the assembler should resolve the issue, not the tokenizer.
This allows us to specify a dot command as a . followed by an identifier, which makes the tokenizer easier.

Resolved issues with newlines contained inside strings.